+++ Light Vox Engine/Physics/RigidBody.h
@@ -108,7 +108,7 @@ namespace Physics
 			const EntityComponents::BoxCollider& two,
 			const glm::vec3& toCenter,
 			EntityComponents::Contacts* contactData,
-			int contactsFound,
+			size_t contactsFound,
 			unsigned best,
 			float penetration);
